backend/requirements.txt
tensorflow-cpu==2.20.0
fastapi>=0.110,<1.0
uvicorn[standard]>=0.24,<1.0
python-multipart>=0.0.9
Pillow>=10.0,<11.0
numpy>2.1
slowapi>=0.1.9